move test to non windwos, we are not making the problem worse
