+++ filebeat/input/filestream/fswatch_test.go
@@ -19,7 +19,6 @@ package filestream
 
 import (
 	"context"
-	"errors"
 	"io/ioutil"
 	"os"
 	"path/filepath"
@@ -27,7 +26,6 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 
 	loginp "github.com/elastic/beats/v7/filebeat/input/filestream/internal/input-logfile"
 	"github.com/elastic/beats/v7/libbeat/common/match"
@@ -118,63 +116,6 @@ func setupFilesForScannerTest(t *testing.T, tmpDir string) {
 		}
 		f.Close()
 	}
-}
-
-func TestFileWatcherRenamedTruncated(t *testing.T) {
-	tmpDir := t.TempDir()
-
-	fs, err := newFileScanner([]string{filepath.Join(tmpDir, "app.log*")}, fileScannerConfig{})
-	if err != nil {
-		t.Fatal(err)
-	}
-	w := fileWatcher{
-		log:          logp.L(),
-		scanner:      fs,
-		events:       make(chan loginp.FSEvent),
-		sameFileFunc: os.SameFile,
-	}
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-	go w.watch(ctx)
-
-	appLogPath := filepath.Join(tmpDir, "app.log")
-	rotatedAppLogPath := filepath.Join(tmpDir, "app.log.1")
-	err = os.WriteFile(appLogPath, []byte("my longer log line"), 0o600)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	evt := w.Event()
-	require.Equal(t, loginp.OpCreate, evt.Op, "new file should be detected")
-	require.Equal(t, "", evt.OldPath, "new file does not have an old path set")
-	require.Equal(t, appLogPath, evt.NewPath, "new file does not have an old path set")
-
-	go w.watch(ctx)
-
-	err = os.Rename(appLogPath, rotatedAppLogPath)
-	if err != nil {
-		t.Fatalf("failed to rotate active file: %v", err)
-	}
-
-	if _, err := os.Stat(appLogPath); !errors.Is(err, os.ErrNotExist) {
-		t.Fatalf("app.log should not exist")
-	}
-
-	err = os.WriteFile(appLogPath, []byte("shorter line"), 0o600)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	evt = w.Event()
-	require.Equal(t, loginp.OpRename, evt.Op, "app.log has been renamed to app.log.1, got: %s old_path=%s new_path=%s", evt.Op.String(), evt.OldPath, evt.NewPath)
-	require.Equal(t, appLogPath, evt.OldPath, "old_path should be set to app.log because of rename")
-	require.Equal(t, rotatedAppLogPath, evt.NewPath, "new_path should be set to app.log.1 because of rename")
-
-	evt = w.Event()
-	require.Equal(t, loginp.OpCreate, evt.Op, "new file app.log should be detected, got: %s for old_path=%s new_path=%s", evt.Op.String(), evt.OldPath, evt.NewPath)
-	require.Equal(t, "", evt.OldPath, "new file should not have an old path set")
-	require.Equal(t, appLogPath, evt.NewPath, "new file should be called app.log")
 }
 
 func TestFileWatchNewDeleteModified(t *testing.T) {
